Revert "Added Mockito and two MockCreation tests"
This reverts commit c4f772208a024d938571cfd0b487d46691b48451.

+++ src/main/java/isi/project/banking/controller/InvestmentController.java
@@ -57,7 +57,7 @@ public class InvestmentController {
 		InvestmentService is = new InvestmentService(em);
 		
 		investment.setOpenDate(new Date());
-		Date closeDate = new Date(investment.getOpenDate().getTime() + (long)investmentDuration*86400000);
+		Date closeDate = new Date((long) (investment.getOpenDate().getTime() + (long)investmentDuration*86400000));
 		investment.setCloseDate(closeDate);
 		is.createInvestment(investment);
 		System.out.println(
